fix: initial animation
